requirements.txt
asgiref==3.6.0
boto3==1.26.114
botocore==1.29.114
certifi==2022.12.7
charset-normalizer==3.1.0
coreapi==2.3.3
coreschema==0.0.4
coverage==7.2.2
Django==4.1.7
django-environ==0.9.0
django-filter==22.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.5
factory-boy==3.2.1
Faker==18.3.1
flake8==5.0.4
flake8-django==1.1.5
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
jmespath==1.0.1
MarkupSafe==2.1.2
mccabe==0.7.0
openapi-codec==1.3.2
packaging==23.0
pdfkit==1.0.0
Pillow==9.4.0
postgres==4.0
psycopg2-binary==2.9.5
psycopg2-pool==1.1
pycodestyle==2.9.1
pyflakes==2.5.0
PyJWT==2.6.0
pyngrok==5.2.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2022.7.1
PyYAML==6.0
reportlab==3.6.12
requests==2.28.2
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
s3transfer==0.6.0
simplejson==3.18.4
six==1.16.0
sqlparse==0.4.3
typing_extensions==4.5.0
uritemplate==4.1.1
urllib3==1.26.15
wkhtmltopdf==0.2
